moved interpolation function call to right before final forcing nc file is written
